crawl_deepseek/pyproject.toml
[project]
name = "auto-crawl-v2"
version = "0.1.0"
description = "Auto crawl DeepSeek with browser-use"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "browser-use",
    "openai",
    "python-dotenv",
    "pydantic>=2.0",
    "httpx",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "mypy",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
